mcalib: replace default RNG by tinymt64
The default libc RNG used was not reentrant and caused problems when intrumenting
programs also using libc rand() and srand().

It is better to use a reentrant RNG in mcalib, also the new RNG is fast and of
much better quality.

Finally we seed it not only with the current time, but also with the pid, to
eventually cope with parallel simulations: we guarantee that simultaneously
launched processes have different starting seeds.