current_mca_interface should not be a pointer
There is a tradeoff with current_mca_interface being a pointer.
When it is a pointer, changing mca backend is cheap but each instrumented op
requires an additional load.

Because we assume that backend changes are less frequent than fop instructions,
we make current_mca_interface be the direct structure address. Now vtable calls
are cheap, but changing mode requires changing all the fields of the vtable.