Fix #96: fix support for bool vectors
Bool vectors are not supported in LLVM, the support for fcmp vector
instructions was broken and we could not return the proper value.

Here we rewrite fcmp vector hooks to return int vectors instead,
libVFCInstrument is modified to perform an appropriate cast when
replacing vector fcmp instructions.

We add a runtime test that checks that the return value of fcmp
instructions is properly handled.