Fix multiple vfc_init calls
The vfcwrapper library constructor may be loaded multiple times.  This
happends for example, when a .so compiled with Verificarlo is loaded
with dlopen into another program also compiled with Verificarlo.

The following hook should ensure that vfc_init is loaded only once.  Is
this code robust? dlopen is thread safe, so this should work.