Integration of vfc_ci source code
Most of the source code/config files have already been modified.

For now, vfc_probes is added as a header (and not as a
precompiled library), which also required to add vfc_hashmap.h
as an accessible header. -> We might want to change this in the
future ?

sigdigits has also been added as a Git submodule, which is downloaded
in the "verificarlo" Makefile rule (there may be a better place to do
this, but it seemed to make sense since this action affects the entire
repository). It is installed as a Python submodule of verificarlo.

vfc_ci is added as a script, but not the ci folder: adding it to the
verificarlo Python does not work since the directory strucure (which is
needed, especially by Bokeh server) is flattened. We still need to
figure out a good way to copy the directory directly into pkgpython
with automake. Copying the folder manually seemed to work perfectly
though, so no other changes may be required to have a first functional
install setup.