Add Python dependencies in Dockerfile and a vfc_ci test
Installing all the new Python dependencies in the Dockerfile should
solve the failing test issues, and result in a Docker image usable with
vfc_ci (including in Github/Gitlab CI).

To make sure that the install of vfc_ci is correct a new test has been
added : it tests the presence of the vfc_ci command itself, as well as
the vfc_probes system.