Write complete test for Fortran interface
Even though the test itself should be correct, it is not expected to work in the current state of this branch, since it relies on vfc_probes_f.mod which still needs to be built wit automake. Moreover, it's unclear how Fortran code should be ignored/built depending on the --with-flang/--without-flang flags.