Integration of CI into master branch (#251)
* Integration of vfc_ci source code

Most of the source code/config files have already been modified.

For now, vfc_probes is added as a header (and not as a
precompiled library), which also required to add vfc_hashmap.h
as an accessible header. -> We might want to change this in the
future ?

sigdigits has also been added as a Git submodule, which is downloaded
in the "verificarlo" Makefile rule (there may be a better place to do
this, but it seemed to make sense since this action affects the entire
repository). It is installed as a Python submodule of verificarlo.

vfc_ci is added as a script, but not the ci folder: adding it to the
verificarlo Python does not work since the directory strucure (which is
needed, especially by Bokeh server) is flattened. We still need to
figure out a good way to copy the directory directly into pkgpython
with automake. Copying the folder manually seemed to work perfectly
though, so no other changes may be required to have a first functional
install setup.

* Copy ci directory into the verificarlo Python module

By using a separate Makefile.am for each ci subdirectory, and
overwriting the pkgpythondir variable, it is possible to copy all the source
files while keeping the same file structure.

* Move most Python scripts into a tools subdirectory

The src/tools subdirectory contains the Python scripts that were
previously in src and postprocessing.

All Makefiles and config files have been edited accordingly.

* Update vfc_ci with changes from dedicated repo

This includes support for multiple repository in one report and some
improvements to the code/comments.

* Replace vfc_put_probe by vfc_probe

* Remove sigdigits module

* Re-initialize sigdigits sumbodules in src/tools/sigdigits

* Add Python dependencies in Dockerfile and a vfc_ci test

Installing all the new Python dependencies in the Dockerfile should
solve the failing test issues, and result in a Docker image usable with
vfc_ci (including in Github/Gitlab CI).

To make sure that the install of vfc_ci is correct a new test has been
added : it tests the presence of the vfc_ci command itself, as well as
the vfc_probes system.

* Add documentation for Verificarlo CI

Also corrected a typo where "05-Postprocessing.md" was linked instead of
"06-Postprocessing.md", and another one with a wrong level of title.

* clang-format vfc_probes and corect permissions on vfc_ci test scripts

These 2 modifications should fix the clang-format-lint and test-docker
tests respectively

* Corrections from verificarlo/verificarlo#238

This includes improvements to the source code of vfc_ci itself
(optimizing some parts of the code, adiing license headers, ...) and to
the installation process (move submodules initialization, ...)

* Fix Dockerfile wit hPython dependencies

REmoved wrong apt packages. Moreover, some Python dependencies were
installed with both apt and pip : this has been removed to use apt if a
package is available, and pip otherwise.

* Custom timerange for run files in report & documentation correction

- It's now possible to specify if some of the most recent run files must
be ignored when generating the report. Coupling this option with a
custom number of files to show makes it possible to select any
chronological range of run files.

- Documentation has been corrected, it used to indicate that runb files
from the las 90 days would appear, instead of the last 100.

* Start adding the Fortran interface to Verificarlo

vfc_probes has been coverted to a shared object library, and the source
code of the Fortran interface has also been added.

Now, it needs to be added to the install process, only when the
--with-flang is specified.

* Update documentation with the Fortran interface

* Write complete test for Fortran interface

Even though the test itself should be correct, it is not expected to work in the current state of this branch, since it relies on vfc_probes_f.mod which still needs to be built wit automake. Moreover, it's unclear how Fortran code should be ignored/built depending on the --with-flang/--without-flang flags.

* Complete installation process ofr vfc_probes_f

The Fortran interface for vfc_probes is now installed when the
--with-flang is specified. If this is the case, the
test_fortran_vfc_probes will also be executed. The interface is
accessible with the vfc_probes_f library as well as its .mod file

* test.sh to test.sh.in in test_fortran_vfc_probes

Now the test.sh script will only be generated if Fortran support is
:x

* Fix Makefile issue in test_fortran_NAS

Co-authored-by: Aurélien Delval <aurelien@babtou-lent.babtou-lent>
Co-authored-by: Pablo de Oliveira Castro <pablo@sifflez.org>