Add asserts and deterministic backends to vfc_ci
Add two important features to vfc_ci :
- support for deterministic backends : different data processing
pipeline for determinstic backend (such as VPREC). This allows to skip
useless computations for these backends (sigma, mu, s, ...) and to have
a different validation system for asserts.
- probes asserts : probes now accept an optional accuracy target. These
probes can be absolute or relative and their satus can be consulted in a
dedicated table in the report.